Moves socket emit function to socket module
