+++ api/routes/bikes.py
@@ -27,7 +27,7 @@ from api.models.models import (
     JsonApiLinks,
     JsonApiResponse,
 )
-from api.socket.socket import socket
+from api.socket.socket import emit_update
 
 router = APIRouter(
     prefix="/v1/bikes",
@@ -197,22 +197,6 @@ async def update_bike(
         data=BikeResource.from_db_model(updated_bike, base_url),
         links=JsonApiLinks(self_link=base_url),
     )
-
-
-async def emit_update(bike_id, updated_bike):
-    """Emits updated bike data to socket."""
-    # TODO: Pydantic model, maybe?
-    print("Emitting")
-    output_data = {
-        "bike_id": bike_id,
-        "last_position": updated_bike.last_position,
-        "city_id": updated_bike.city_id,
-        "battery_lvl": updated_bike.battery_lvl,
-        "is_available": updated_bike.is_available,
-        "meta_data": updated_bike.meta_data,
-    }
-    await socket.emit("bike_update", data=output_data, room="bike_updates")
-    return
 
 
 @router.delete("/{bike_id}", status_code=status.HTTP_204_NO_CONTENT)
